requirements.txt
google-generativeai
pandas
numpy
watchdog
